requirements.txt
backoff
beautifulsoup4
openai
openpyxl
pandas
python-dotenv
python_dateutil
pytz
selenium
tzlocal